+++ src/components/user-list/user-list.ts
@@ -159,7 +159,7 @@ export class UserListComponent implements OnInit {
 
     cache(): void {
         console.log('Load cache', this.params);
-        this.provider.findCache(this.params).then(_data => {
+        this.provider.list(this.params).then(_data => {
             console.log('cache', _data);
             if (_data.length) {
                 _.sortBy(_data, 'createdAt').reverse().map(item => this.data.push(item));