+++ src/pages/tab-search/tab-search.ts
@@ -1,31 +1,32 @@
-import {Component} from "@angular/core";
-import {Events} from "ionic-angular";
+import {Component, ViewChild} from "@angular/core";
+import {Events, Content} from "ionic-angular";
 import {AnalyticsProvider} from "../../providers/analytics.provider";
 
 @Component({
-    selector   : 'page-tab-search',
+    selector:    'page-tab-search',
     templateUrl: 'tab-search.html'
 })
 export class TabSearchPage {
 
-    search: string      = '';
+    @ViewChild(Content) content: Content;
+
+    search: string = '';
     placeholder: string = 'Search';
     username: string;
-    type: string        = 'photo';
+    type: string = 'photo';
     profile: any;
-    moreItem: boolean   = false;
-    eventName: string   = 'search';
+    moreItem: boolean = false;
+    eventName: string = 'search';
 
     params = {
-        limit    : 12,
-        page     : 1,
+        limit:     12,
+        page:      1,
         privacity: 'public',
-        search   : '',
+        search:    '',
     };
 
     constructor(private analytics: AnalyticsProvider,
-                private events: Events
-    ) {
+                private events: Events) {
         // Google Analytics
         this.analytics.view('TabSearchPage');
 
@@ -64,15 +65,24 @@ export class TabSearchPage {
     }
 
     public onReload(): void {
-        this.params.page   = 1;
+        this.params.page = 1;
         this.params.search = "";
         this.sendParams();
+        this.scrollToTop();
     }
 
     // Search
     public doSearch(): void {
         this.params.search = this.search;
-        this.params.page   = 1;
+        this.params.page = 1;
         this.sendParams();
+    }
+
+    scrollToTop(): void {
+        setTimeout(() => {
+            if (this.content.scrollTop) {
+                this.content.scrollToTop(100)
+            }
+        }, 100);
     }
 }