+++ src/providers/gallery-album.provider.ts
@@ -1,5 +1,4 @@
 import {Injectable} from "@angular/core";
-import * as PouchDB from "pouchdb";
 declare var Parse: any;
 
 @Injectable()
@@ -17,12 +16,9 @@ export class GalleryAlbumProvider {
 
     private _ParseObject: any = Parse.Object.extend('GalleryAlbum', {});
 
-    db: any;
     data: any[] = [];
 
     constructor() {
-
-        this.db = new PouchDB('GalleryAlbum', {auto_compaction: true});
 
         this._fields.map(field => {
             Object.defineProperty(this._ParseObject.prototype, field, {
@@ -45,31 +41,6 @@ export class GalleryAlbumProvider {
 
     find(params: any): Promise<any> {
         return Parse.Cloud.run('listAlbum', params);
-    }
-
-    cleanCache(): Promise<any> {
-        this.data = [];
-        return new Promise(resolve => {
-            this.db
-                .allDocs({include_docs: true})
-                .then(result => Promise.all(result.rows.map(row => this.db.remove(row.doc))).then(resolve));
-        });
-    }
-
-
-    findCache(params?:any): Promise<any> {
-        return new Promise(resolve => {
-            if (this.data.length > 0) {
-                resolve(this.data)
-            } else {
-                this.db.allDocs({include_docs: true}).then(data => {
-                    if (data.total_rows) {
-                        this.data = data.rows.map(row => row.doc);
-                    }
-                    resolve(this.data);
-                })
-            }
-        });
     }
 
     photo(params: any) {