+++ src/providers/gallery.provider.ts
@@ -2,14 +2,9 @@ import {Injectable} from "@angular/core";
 import {IParamsLocation} from "../models/parse.params.location.model";
 import {IGallery} from "../models/gallery.model";
 import {IParams} from "../models/parse.params.model";
-import * as PouchDB from "pouchdb";
-import _ from "underscore";
 
 declare const Parse: any;
-declare const require: any;
 
-//PouchDB.debug.enable('*');
-//PouchDB.debug.disable();
 
 @Injectable()
 export class GalleryProvider {
@@ -34,12 +29,9 @@ export class GalleryProvider {
 
     private _ParseObject: any = Parse.Object.extend('Gallery', {});
 
-    db: any;
     data: any[] = [];
 
     constructor() {
-
-        this.db = new PouchDB('Gallery', {auto_compaction: true});
 
         this._fields.map(field => {
             Object.defineProperty(this._ParseObject.prototype, field, {
@@ -95,56 +87,10 @@ export class GalleryProvider {
         return Parse.Cloud.run('updateGallery', params);
     }
 
-    public cleanCache(): Promise<any> {
-        this.data = [];
-        return new Promise((resolve, reject) => {
-            this.db
-                .allDocs({include_docs: true})
-                .then(result => Promise.all(result.rows.map(row => this.db.remove(row.doc))))
-                .then(resolve)
-                .catch(reject);
-        });
-    }
 
     public feed(params: IParams): Promise<any> {
         return Parse.Cloud.run('feedGallery', params);
     }
-
-    public feedCache(data: any): Promise<any> {
-        let db = new PouchDB('Gallery', {auto_compaction: true});
-        return Promise.all(data.map(item => db.put(item)));
-    }
-
-    public findCache(params: IParams): Promise<any> {
-        return new Promise(resolve => {
-            this.db.allDocs({include_docs: true}).then(data => {
-                console.log(data);
-                this.data = [];
-                if (data.total_rows) {
-                    data.rows.map(row => {
-                        //let doc = JSON.stringify(row.doc.data);
-                        row.doc.createdAt = new Date(row.doc.createdAt);
-                        this.data.push(row.doc);
-                    });
-
-                    if (params.username) {
-                        let _data = _.find(this.data, {username: params.username});
-                        console.log('cache username', _data);
-                        resolve(_data);
-                    } else {
-                        console.log('cache', this.data);
-                        resolve(this.data);
-                    }
-                } else {
-                    resolve(this.data);
-                }
-            })
-        });
-    }
-
-    //feed(params: any){
-    //    return Parse.Cloud.run('feedGallery', params);
-    //}
 
     comments(params) {
         return Parse.Cloud.run('commentGallery', params);
@@ -159,9 +105,6 @@ export class GalleryProvider {
         return new Parse.Query(this._ParseObject).get(objectId);
     }
 
-    getCache(objectId: string): Promise<any> {
-        return this.db.get(objectId);
-    }
 
     createGallery(form:any){
         return Parse.Cloud.run('createGallery', form);